Allow selection of any project view including ComopositeView with projection
List the views on the project in the EditTableForm component rather than
just the default Sparql and ES indices. If the view is a CompositeView
then we show a dropdown to select the projection. The endpoint queried
will depend on the view type.

Currently, ES queries not showing any data in table - how do we do this?
We also need to support searching all projections of a particular type
and aggregate views too.